+++ __tests__/unit/Shimmer.test.tsx
@@ -1,4 +1,4 @@
-import { Shimmer } from "../../components/shimmer/index";
+import { Shimmer } from "@/components/shimmer/shimmer";
 import { ListShimmer } from "@/components/shimmer/ListShimmer";
 import { cleanup, render, screen } from "@testing-library/react";
 import { afterEach, describe, expect, test, vi } from "vitest";
